Overwrite signature
